Update README.md
Tkinter is a standard Python library for creating graphical user interfaces. It provides a wide range of widgets such as labels, buttons, and entry fields, which can be used to create a variety of different user interfaces. This project comprised of username entry, password entry and Buttons like Login Button, Forgot Password Button and Sign Up Button.

Another thing to look out for in this project is the eye toggle buttons which will be used to hide and show password the user enters inside the Password Entry Widget